Phase 2

+++ routes/review.js
@@ -5,6 +5,7 @@ const ExpressError = require("../utils/ExpressError");
 const {reviewSchema} = require("../schema");
 const Listing = require("../models/listing");
 const Review = require("../models/review");
+const {isloggedin} = require("../middleware");
 
 const validateReview = (req,res,next)=>{
     let {error}= reviewSchema.validate(req.body);
@@ -17,10 +18,12 @@ const validateReview = (req,res,next)=>{
     }
 }
 
-router.post("/",validateReview,wrapAsync(async(req,res)=>{
+router.post("/",isloggedin, validateReview,wrapAsync(async(req,res)=>{
     let {id} = req.params;
     let listing = await Listing.findById(id);
     let newReview = new Review(req.body.review);
+    newReview.author = req.user._id;
+    console.log(newReview.author);
     await newReview.save();
 
     listing.reviews.push(newReview);
@@ -33,6 +36,11 @@ router.post("/",validateReview,wrapAsync(async(req,res)=>{
 //delete review route
 router.delete("/:reviewId", wrapAsync(async(req,res)=>{
     let{id, reviewId} = req.params;
+    const review = await Review.findById(reviewId);
+    if(!review.author.equals(res.locals.currUser.id)){
+        req.flash("error","You are not the author of this review");
+        return res.redirect(`/listings/${id}`);
+    }
     await Listing.findByIdAndUpdate(id,{$pull : {reviews : reviewId}});
     await Review.findByIdAndDelete(reviewId);
     req.flash("success","Review deleted")
